Maps and Kits
Fixed all the bugs from yesterday's update.

Kits are now in-progress! The main base for kits have been implemented. Archer only exists as of now, but the rest of the kits are coming soon! 

HB's smoke is still broken and seems to cause a bit of lag, so I'll look into that

Updated the shard timeout to 5 mins like in the real game, along with changing the shard pickup slowness from 20 seconds to 30 seconds

Updated the waiting actionbars to fit the actual game's a bit better.

Redis is now a requirement. Details must be put in the plugin's config